requirements.txt
streamlit
pymongo
pyyaml
linkml